gemini-mcp-server/pyproject.toml
[project]
name = "gemini-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.68.0",
    "mcp[cli]>=1.26.0",
]
